[#90] cleaned Utils in ui folder
